test_recorder/requirements.txt
playwright==1.48.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
openpyxl==3.1.2
python-docx==1.1.0
pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
sqlite-utils==3.35.2
loguru==0.7.2
asyncio-mqtt==0.16.1
websockets==12.0
watchdog>=4.0.0 